examples/applications/NII/requirement.txt
boto3>=1.26.126
numpy ==1.22
python-dotenv>=1.0.0
tensorflow==2.12.0
flask
flask_restful
paho-mqtt
redis